% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{calcAFD}
\alias{calcAFD}
\title{Calculate allele frequency differences by block loess regression}
\usage{
calcAFD(
  phasedCounts,
  experiment.name = "",
  sample.size = 250000,
  sel.strength = 0.5,
  bin.width = 500,
  eff.length = 600,
  gmap = NULL,
  vcf.cross = NULL,
  uchr = paste0("chr", as.roman(1:16))
)
}
\arguments{
\item{phasedCounts}{data frame with columns ID, p1, p2, which are marker ID, counts for parent 1 and counts for parent 2}

\item{experiment.name}{column name in vcf for experiment of interest}

\item{sample.size}{sample size of unselected population}

\item{sel.strength}{fraction of population selected}

\item{bin.width}{collapse counts to bins every 500bp (default) or user specificied physical marker positions (optional)}

\item{eff.length}{effective number of indepedent tests across the genome}

\item{gmap}{genetic map object (optional)}

\item{vcf.cross}{vcfR object containing the two parent variant calls (optional)}

\item{uchr}{a vector of chromosome names (optional)}
}
\value{
data.frame of smoothed results with SE
}
\description{
Calculate allele frequency differences by block loess regression
}
