% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{calcContrastStats}
\alias{calcContrastStats}
\title{Calculate allele frequency between tails, or tail and control. Convert to Z, p-value, and LOD.}
\usage{
calcContrastStats(results, L = "meta_high", R = "meta_low")
}
\arguments{
\item{results}{list of data frames of results to compare, should contain column names of betas and se with prefix L and R}

\item{L}{prefix for column name containing beta (L) and SE L.se, or a numeric constant (e.g. 0.5) to test R against}

\item{R}{prefix for column name containing beta (R) and SE R.se, a numeric constant (e.g. 0.5) to test L against, or NULL for single-column z-test}
}
\value{
combined.results meta analysis Betas and SEs added to results merged across the replicates
}
\description{
Calculate allele frequency between tails, or tail and control. Convert to Z, p-value, and LOD.
}
