% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{calcMetaAFD}
\alias{calcMetaAFD}
\title{Use inverse variance weighting to combine information across replicates}
\usage{
calcMetaAFD(results, meta_name = "meta")
}
\arguments{
\item{results}{data frame of results, should contain smoothed allele-frequency differences (afd) and SEs (afd.se)}

\item{meta_name}{specify name of meta analysis beta and se}
}
\value{
combined.results meta analysis Betas and SEs added to results merged across the replicates
}
\description{
Use inverse variance weighting to combine information across replicates
}
