% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{doBlockLoess}
\alias{doBlockLoess}
\title{Do loess regression of p1/(p1+p2) per block, in user defined blocks}
\usage{
doBlockLoess(p1, p2, pos, bin.width = 500)
}
\arguments{
\item{p1}{counts for parent 1 at each position}

\item{p2}{counts for parent 2 at each position}

\item{pos}{positions for each marker}

\item{bin.width}{collapse counts to bins every 500bp (default) or user specificied physical marker positions (optional)}
}
\value{
imputed fitted allele frequency delta at each pos
}
\description{
Do loess regression of p1/(p1+p2) per block, in user defined blocks
}
