% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findIntervalArmadillo}
\alias{findIntervalArmadillo}
\title{Find interval implementation in armadillo}
\usage{
findIntervalArmadillo(x, breaks)
}
\arguments{
\item{x}{the cutoff points to be found}

\item{breaks}{the vector to look for.}
}
\description{
This takes the fast findInterval2 implementation from Hadley Wickham (http://adv-r.had.co.nz/Rcpp.html) and just replaces everything with RcppArmadillo objects. Since smoothQTL is implemented in armadillo this saves some conversions
}
