% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{generateLen}
\alias{generateLen}
\title{Title Internal funcion to turn values into NA based on filter}
\usage{
generateLen(vec, filt)
}
\arguments{
\item{vec}{vector}

\item{filt}{filter vector (boolean)}
}
\value{
output with filtered values turned to NA. This function is basic, only convenience to avoid repeating code.
}
\description{
Title Internal funcion to turn values into NA based on filter
}
