% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{getBiallelicCounts}
\alias{getBiallelicCounts}
\title{Extract reference and alt counts for a given experiment from a VCF}
\usage{
getBiallelicCounts(
  experiment.vcf,
  exp.name,
  uchr = paste0("chr", as.roman(1:16))
)
}
\arguments{
\item{experiment.vcf}{vcfR object for the biallelic variants segregating for a given cross and allelic depths for reference and alt alleles}

\item{exp.name}{column name in vcf for experiment of interest}

\item{uchr}{vector of chromosome names in sorted desired order (optional)}
}
\value{
data.frame of variant ID, ref count, and alt count
}
\description{
Extract reference and alt counts for a given experiment from a VCF
}
