% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{getGmapPositions}
\alias{getGmapPositions}
\title{Given a reference genetic map and physical positions of markers, return imputed genetic positions}
\usage{
getGmapPositions(vcf.cross, gmap, uchr)
}
\arguments{
\item{vcf.cross}{vcfR object for the biallelic variants segregating for a given cross}

\item{gmap}{a genetic map object}

\item{uchr}{a vector of chromosome names}
}
\value{
a list of cM positions for given physical positions for markers on each chromosome
}
\description{
Given a reference genetic map and physical positions of markers, return imputed genetic positions
}
