% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{getIDSorted}
\alias{getIDSorted}
\title{Given a vcfR object, get marker IDs sorted in intended chromosome order}
\usage{
getIDSorted(vcf.cross, uchr)
}
\arguments{
\item{vcf.cross}{vcfR object for the biallelic variants segregating for a given cross}

\item{uchr}{vector of chromosome names in intended order}
}
\value{
vector of marker IDs
}
\description{
Given a vcfR object, get marker IDs sorted in intended chromosome order
}
