% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{getPars}
\alias{getPars}
\title{Title A robust fit for the log normal distribution. This is just a wrapper for the fitdstnRob function from robust.}
\usage{
getPars(Gsmoothed)
}
\arguments{
\item{Gsmoothed}{Smoothed G statistics}
}
\value{
See fitdstnRob function in package robust. This function just returns the estimate (the fitted mean and sd).
}
\description{
Title A robust fit for the log normal distribution. This is just a wrapper for the fitdstnRob function from robust.
}
