% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{getq}
\alias{getq}
\title{Title get allele frequences for allele A}
\usage{
getq(n1, n3)
}
\arguments{
\item{n1}{counts for allele A}

\item{n3}{counts for allele B}
}
\value{
Frequencies of allele A. Note, to prevent a catastrophe when calculating G statistic, frequencies of 1 and 0 are converted to 0.99 and 0.01.
}
\description{
Title get allele frequences for allele A
}
