% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{getSampleSizes}
\alias{getSampleSizes}
\title{Esimate sample sizes either given total pop (unselected) sample size and selection strengths
or effective depths if sample size is constrained by sequencing depth}
\usage{
getSampleSizes(results.data, sample.size, sel.high, sel.low, eff.length = 600)
}
\arguments{
\item{results.data}{data frame of results}

\item{sample.size}{size of unselected population}

\item{sel.high}{fraction of population in high tail}

\item{sel.low}{fraction of population in low tail}

\item{eff.length}{(default=600) effective number of tests across the genome}
}
\value{
list of effective sample size estimates
}
\description{
Esimate sample sizes either given total pop (unselected) sample size and selection strengths
or effective depths if sample size is constrained by sequencing depth
}
