% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gqtl}
\alias{gqtl}
\title{Title Run X-QTL analysis}
\usage{
gqtl(
  bulkA,
  bulkB = NULL,
  geneticMap,
  splitVar,
  getDistortion = T,
  kern = function(d) {     (1 - d^3)^3/sum((1 - d^3)^3) },
  W = 25,
  lim = 0.99
)
}
\arguments{
\item{bulkA}{A two column matrix with the allele counts in the low/control group}

\item{bulkB}{A two column matrix with the allele counts in the high/treatment group}

\item{geneticMap}{A vector of the genetic distances. All chromosomes are combined into one vector, and within each chromosome the values should monotonically increase. The function is compatible with the genetic map vector produced by r/qtl. has to be the same length as the row number of bulkA and bulkB}

\item{splitVar}{A vector used to split the genetic map and the allele counts to chromosomes. Has to be the same length as the row number of bulkA and bulkB}

\item{getDistortion}{Should segregration distortion be calculated from the data? If FALSE, X-QTL will run assuming baseline allele frequencies are all 0.5.}

\item{kern}{The kernel function for the smoother. This is currently a placeholder since the function uses a C++ implementation of the smoother with the kernel baked in.}

\item{W}{The window size for the smoother. When smoothing the value for an SNV, a window of size W will be taken around it. See the Magwene et al 2011 paper for discussion of values. Defaults to 25, the low bound of the values they recommend}

\item{lim}{A filter for SNVs which show distortion. This is calculated based on the low/control group. The G statistic becomes highly unstable when considering SNVs that are nearly fixed, so they are filtered out.}
}
\value{
A list with the following values:
GstatA: The G statistic for the control/low group (compared to a baseline of 0.5/0.5). Can be used to determine segregation distortion in the baseline
GstatB: The G statistic for the high/treatment group (compared to the control group). This is the statistic used for the X-QTL
GsmoothA: The smoothed G statistic for the control/low group. 
GsmoothB: The smoothed G statistic for the treatment/high group. Used directly to calculate P-values.
PvalA: P-values for segregation distortion in the baseline. Beware! If large parts of the genome show distortion (as in C. elegans) This will not work, since it derives the null distribution robustly from the values themselves.
PvalB: XQTL P-values.
qSmoothA: The smoothed allele frequencies in the control/low group. Can be useful to observe segregation distortion in the baseline
qSmoothB: The smoothed allele frequences in the treatment/high group. Together with qSmoothA, can be used to determine direction of observed QTL
geneticMap: The genetic map supplied to the function. This is returned for convenience/sanity check
splitVar: The chromosome vector supplied to the function. This is returned for convenience/sanity check
}
\description{
Title Run X-QTL analysis
}
