% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gqtlDF}
\alias{gqtlDF}
\title{Title Transform XQTL list to data frame}
\usage{
gqtlDF(res)
}
\arguments{
\item{res}{Results from the gqtl function}
}
\value{
Convenience function that simply converts the list returned by gqtl to data frame
}
\description{
Title Transform XQTL list to data frame
}
