% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{Gsmooth}
\alias{Gsmooth}
\title{Title Depracated slow R version of the G statistic smoother. Can be useful to test kernels}
\usage{
Gsmooth(
  G,
  kern = function(d) {     (1 - d^3)^3/sum((1 - d^3)^3) },
  map,
  W = 25
)
}
\arguments{
\item{G}{Raw G statistic}

\item{kern}{The Kernel function to be used. Should be a function that gets a single parameter (d - distance from variant) and returns the transformed distance.}

\item{map}{The genetic map vector (see gqtl)}

\item{W}{The window for smoothing (see gqtl)}
}
\value{
Vector of smoothed G values
}
\description{
Title Depracated slow R version of the G statistic smoother. Can be useful to test kernels
}
