% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{Gstat}
\alias{Gstat}
\title{Title Calculate G statistic}
\usage{
Gstat(n2, n4, n1 = NULL, n3 = NULL, q = 0.5)
}
\arguments{
\item{n2}{Counts of allele A and group 1}

\item{n4}{Counts of allele B and group 1}

\item{n1}{Counts of allele A and group 2}

\item{n3}{Counts of allele B and group 2}

\item{q}{The baseline allele frequencies. If no n1 and n3 are given, q is used to simulate allele counts using the same total counts as n2+n4.}
}
\value{
A vector of G statistic for each variant
}
\description{
Title Calculate G statistic
}
