% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{jitterGmapVector}
\alias{jitterGmapVector}
\title{Jitter a genetic map so map positions don't completely overlap}
\usage{
jitterGmapVector(themap, amount = 1e-06)
}
\arguments{
\item{themap}{a list of vectors of marker positions in cM}

\item{amount}{amount of jitter}
}
\value{
a list of vectors of jittered marker positions
}
\description{
Jitter a genetic map so map positions don't completely overlap
}
