% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{LODToPval}
\alias{LODToPval}
\title{Convert LOD to p-value}
\usage{
LODToPval(x)
}
\arguments{
\item{x}{LOD score}
}
\value{
p-value
}
\description{
Convert LOD to LRS, and then assume LRS is χ2 distributed with 1 d.f.
}
