% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_analysis.R
\name{makeCountTablesGATK}
\alias{makeCountTablesGATK}
\title{Subset GATK table for known segregating variants and phase given parental genotypes}
\usage{
makeCountTablesGATK(sample_dir, allele_counts, p.names, vcf, gt, gmap)
}
\arguments{
\item{sample_dir}{sample directory}

\item{allele_counts}{name of file}

\item{p.names}{expected parents}

\item{vcf}{vcfR object}

\item{gt}{genotype calls from vcfR object}

\item{gmap}{genetic map}
}
\value{
data.frame with original and phased counts
}
\description{
Subset GATK table for known segregating variants and phase given parental genotypes
}
