% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_analysis.R
\name{phaseBiparental}
\alias{phaseBiparental}
\title{Phase reference and alt counts given the ref/alt calls for one of the parents, using AlphaSimR objects}
\usage{
phaseBiparental(df, p1.name, founderPop, genMap)
}
\arguments{
\item{df}{data.frame output from getBiallelicCounts() should contain the columns: ID,ref,alt}

\item{p1.name}{name of parent to be designated parent 1 in vcf.cross (name must exist in vcf.cross)}

\item{founderPop}{\subsection{}{
}}
}
\value{
data.frame of variant ID, ref count, and alt count phased
}
\description{
Phase reference and alt counts given the ref/alt calls for one of the parents, using AlphaSimR objects
}
