% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{phaseCounts}
\alias{phaseCounts}
\title{Phase reference and alt counts given the ref/alt calls for one of the parents}
\usage{
phaseCounts(vcf.cross, p1.name, df, uchr = paste0("chr", as.roman(1:16)))
}
\arguments{
\item{vcf.cross}{vcfR object for the biallelic variants segregating for a given cross}

\item{p1.name}{name of parent to be designated parent 1 in vcf.cross (name must exist in vcf.cross)}

\item{df}{data.frame output from getBiallelicCounts() should contain the columns: ID,ref,alt}

\item{uchr}{vector of chromosome names in sorted desired order (optional)}
}
\value{
data.frame of variant ID, ref count, and alt count phased
}
\description{
Phase reference and alt counts given the ref/alt calls for one of the parents
}
