% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{PvalToLOD}
\alias{PvalToLOD}
\title{Convert p-value to LOD score}
\usage{
PvalToLOD(x)
}
\arguments{
\item{x}{p-value}
}
\value{
LOD score
}
\description{
Convert p-value to LOD score, inverse of LODToPval
}
