% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{RtoLOD}
\alias{RtoLOD}
\title{Convert pearson R to LOD score}
\usage{
RtoLOD(n, r)
}
\arguments{
\item{n}{sample size}

\item{r}{pearson r}
}
\value{
LOD score
}
\description{
Convert pearson R to LOD score
}
