% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{simBayesC}
\alias{simBayesC}
\title{Simulate phenotypes for haploid prograny for an XSnpMatrix with reference-based biallelic coding coming from a BayesC genetic architecture}
\usage{
simBayesC(geno.matrix, target.size = 0.1, h2 = 0.4)
}
\arguments{
\item{geno.matrix}{XSnpMatrix}

\item{target.size}{fraction of markers with effects sampled from N(0,h^2)}

\item{h2}{total additive heritability}
}
\value{
y vector of simulated phenotype, scaled so total phenotypic variance is 1
}
\description{
Simulate phenotypes for haploid prograny for an XSnpMatrix with reference-based biallelic coding coming from a BayesC genetic architecture
}
