% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{simHaploidSegsFN}
\alias{simHaploidSegsFN}
\title{Simulate genotypes for haploid progeny for a given cross given known biallelic variants and a genetic map}
\usage{
simHaploidSegsFN(
  vcf.cross,
  gmap,
  nsegs,
  uchr = paste0("chr", as.roman(1:16)),
  ngenerations = 2
)
}
\arguments{
\item{vcf.cross}{vcfR object for the biallelic variants segregating for a given cross}

\item{gmap}{a genetic map object}

\item{nsegs}{number of desired haploid progeny}

\item{uchr}{a vector of chromosome names}

\item{ngenerations}{(default=2) number of generations of intercrossing, assuming only progeny from most recent generation mate etc}
}
\value{
XSnpMatrix snpStats matrix object of 0/1 (relative to reference) coded genotypes of progeny
}
\description{
Simulate genotypes for haploid progeny for a given cross given known biallelic variants and a genetic map
}
