% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{simPhenotypesForRefPop}
\alias{simPhenotypesForRefPop}
\title{Simulate phenotypes for haploid progeny for an XSnpMatrix with reference-based biallelic coding}
\usage{
simPhenotypesForRefPop(sx, h2 = 0.1, nQTL = 1)
}
\arguments{
\item{sx}{XSnpMatrix}

\item{h2}{total additive heritability}

\item{nQTL}{total number of QTL}
}
\value{
y vector of simulated phenotype, scaled so total phenotypic variance is 1
}
\description{
Simulate a number of QTL with equal effects (and random sign) with known summed total h^2
}
