% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_sim.R
\name{simXQTLExperiment}
\alias{simXQTLExperiment}
\title{Simulate an xQTL experiment with reference-based 0/1 counts for ref and alt alleles}
\usage{
simXQTLExperiment(
  y,
  geno.matrix,
  vcf.cross,
  sel.low = 0.1,
  sel.high = 0.1,
  depth.low = 50,
  depth.high = 50,
  vcf.out = NULL
)
}
\arguments{
\item{y}{vector of simulated phenotype, scaled so total phenotypic variance is 1}

\item{geno.matrix}{XSnpMatrix}

\item{vcf.cross}{vcfR object for the biallelic variants segregating for a given cross}

\item{sel.low}{fraction of population selected for low tail}

\item{sel.high}{fraction of population selected for high tail}

\item{depth.low}{average depth per variant site for low tail (summed over both alleles)}

\item{depth.high}{average depth per variant site for low tail (summed over both alleles)}

\item{vcf.out}{filename of simulated vcf file output}
}
\value{
expected allele frequencies for low and high tail
}
\description{
Simulate an xQTL experiment with reference-based 0/1 counts for ref and alt alleles
}
