% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smoothQTL}
\alias{smoothQTL}
\title{Calculate smoothed G statistic}
\usage{
smoothQTL(G, Map, W)
}
\arguments{
\item{G}{the G statistic}

\item{Map}{the genetic map (see gqtl function)}

\item{W}{the window size for the smoothing (see gqtl function)}
}
\description{
This function implements the smoothing procedure outlined in Magwene et al 2011.
}
