\name{xQTLstats-package}
\alias{xQTLstats-package}
\alias{xQTLstats}
\docType{package}
\title{
\packageTitle{xQTLstats}
}
\description{
\packageDescription{xQTLstats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xQTLstats}
\packageIndices{xQTLstats}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{xQTLstats}

Maintainer: \packageMaintainer{xQTLstats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }

