% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xQTL_stats.R
\name{plotIndividualExperiment}
\alias{plotIndividualExperiment}
\title{Plot individual experiment results}
\usage{
plotIndividualExperiment(results, suffix, simulatedQTL = NULL)
}
\arguments{
\item{results}{data.frame of experiment results}

\item{suffix}{of experiment name}

\item{simulatedQTL}{data frame of chromosome, and position of simulated QTL (optional)}
}
\value{
ggplot object of experiment results
}
\description{
Plot individual experiment results
}
